+++ README.md
@@ -16,3 +16,4 @@ setting up jest requires config
 npm i -D jest @types/jest ts-jest msw whatwg-fetch @testing-library/react @testing-library/jest-dom swr
 ts-node
 jsdom
+- embla-carousel-react
